File reads and writes now update file times.
As part of this, ByteStoreFileChannel has been renamed to JimfsFileChannel and moved out of the bytestore package.
It now stores the File as well as the ByteStore so it can update file times. ByteStoreInputStream and ByteStoreOutputStream
have been removed in favor of just delegating to the FileChannel once again, for simplicity. ByteStore has been simplified
to just have methods dealing with ByteBuffers once again.