Add support for viewing in-memory ZIP/JAR files as FileSystems using FileSystems.newFileSystem(Path, ClassLoader).
This is done by creating a "jar:jimfs://..." URI for the path and then passing it off to be handled by the
built in ZipFileSystemProvider. Works great as long as you just want to read the file; writing entries works
but fails when it tries to sync the changes to the file system, as for some reason ZipFileSystem is
hardwired to the default file system for that in a weird way.

This change also slightly improves JimfsPath.toUri() to more or less work, but it still could use
improvement.