Add Disk, an abstract in-memory shared file storage, and DiskByteStore, a ByteStore backed by a Disk.
Also add two implementations of Disk: SingleArrayDisk and MultiArrayDisk.
SingleArrayDisk is backed by a single size-doubling array.
MultiArrayDisk is backed by multiple fixed size arrays that are allocated as needed.

FileContent now has a delete() method, to allow a DiskByteStore to free the blocks allocated to it
for reuse by the Disk.