Greatly simplify attribute handling.
The change gets rid of the Attribute class and the associated AbstractAttributeProvider,
instead leaving each AttributeProvider to handle some of the type and settability checking
that was handled by AttributeService before. This turns out to be easy to do in each
AttributeProvider implementation and cuts down on a lot of complication.

This also gets rid of the AttributeViews configuration class in favor of an
AttributeConfiguration class. Attribute views are now configured by simply giving a
list of String view identifiers for the views you want the file system to support.
Custom providers can be added separately. Finally, overriding the default values for
attributes is now handled by simply mapping the String attribute identifier
(e.g "posix:group") to a default value in the AttributeConfiguration, allowing the
AttributeProvider implementations to be singletons.