Get rid of FileContent interface, changing all implementations to subclasses of File.
While logically it seemed to make sense for the content of each file to be a separate object, it led to awkward code particularly when dealing with directories (since the content needed a reference to the File that contained it) and regular file streams/channels, which needed both a reference to the File and the content anyway.

- ByteStore renamed to RegularFile.
- DirectoryTable renamed to Directory.
- DirectoryEntryMap inlined in to Directory.
- SymbolicLink class added, which just contains a JimfsPath target.
- Various code simplified to accept or return specific File subclasses.
- Also re-add a package-info.java file, mostly just for @ParametersAreNonNullByDefault (for whatever good that does).

This also has the effect of further reducing the number of objects per file in the file system. For example, where a directory was previously a File, DirectoryTable and DirectoryEntryMap, it's now just a Directory.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=60676208