Change directories to *not* unset their link to the parent directory when they are unlinked.
Instead, when the directory is then linked in a new directory (as in a move), the parent link is overwritten with a new entry. When the directory is deleted, the parent link remains as long as the file object hangs around (as long as a SecureDirectoryStream is open to it, or for the lifetime of the FileSystem if the working directory was deleted.).

This matches the behavior observed running in Linux, the only environment I've tried that supports SecureDirectoryStream.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=60680760