Add a method for getting the Configuration that the Jimfs.newFileSystem() overloads that don't take a Configuration use.
This is primarily for allowing the user to create a modified version of that default configuration, which currently isn't possible without duplicating the logic to get the configuration.

Also remove an unused constant. It's public, but it seems unlikely anyone's using it.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=112255563