Throw UOE when setting unsupported attributes
This commit modifies the behavior when setting unsupported attributes to
throw an UnsupportedOperationException instead of an
IllegalArgumentException. This is to make
JimfsFileSystemProvider#createDirectory and
JimfsFileSystemProvider#newByteChannel behave consistently with the
Javadocs for FileSystemProvider#createDirectory and
FileSystemProvider#newByteChannel which require an
UnsupportedOperationException to be thrown when the array of attributes
contains attributes which can not be set. In particular, these methods
are invoked via Files#createTempDirectory and Files#createFile which
have the same requirement regarding throwing an
UnsupportedOperationException when the array of attributes contains
attributes which can not be set. Most notably, this causes a discrepancy
between the handling of Jimfs when configured to act like a Windows
filesystem versus the behavior of sun.nio.fs.WindowsFileSystem where the
former will throw an IllegalArgumentException if an attempt is made to
set a POSIX attribute but the latter will throw an
UnsupportedOperationException (consistent with the Javadocs).