Ensure streams that encapsulate a closeable resource are closed
This cleanup is being performed in preparation for enabling a
compiler error.

From the javadoc for Files.list:

The returned stream encapsulates a DirectoryStream. If timely disposal
of file system resources is required, the try-with-resources construct
should be used to ensure that the stream's close method is invoked
after the stream operations are completed.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=180930883