Update maven-javadoc-plugin to 3.0.0.
This is necessary to build with JDK9.
(You can still build *for* older versions of Java to maintain compatibility; this is just to work with build tools from JDK9, as will become the Google default.)

As part of that, migrate from additionalparam, which was deprecated and removed last year, to additionalOptions.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=182543366