Fix an issue with Files.newOutputStream for Jimfs where it wouldn't truncate the file if the TRUNCATE_EXISTING option was explicitly provided and the WRITE option wasn't.
The code that handles TRUNCATE_EXISTING requires that WRITE also be present for it to do anything (because it's documented that it's ignored otherwise), but newOutputStream should always add the WRITE option if the user doesn't provide it because an OutputStream is obviously for writing. (This matches the default behavior of FileSystemProvider.newOutputStream as well.)

Fixes #77

RELNOTES=Fixed an issue where `Files.newOutputStream` wouldn't truncate the file if the `TRUNCATE_EXISTING` option was explicitly provided and the `WRITE` option wasn't.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=237818016