Migrate Truth subjects from a manual "test and fail" approach to one using Subject.check(...).
Before:
  if (actual().foo() != expected) {
    fail("has foo", expected);
  }

After:
  check("foo()").that(actual().foo()).isEqualTo(expected);

This CL changes the format of the failure message. The new message will typically contain the same information as the old, but if some is missing, let us know, and consider adding a test.

The overload of check(...) used by this CL was added in Truth 0.40.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=244408001