Migrate jimfs from `@NullableDecl` to `@Nullable`.
This was prompted by the CI failure of https://github.com/google/jimfs/pull/240, in which `maven-javadoc-plugin` had trouble finding `@NullableDecl`. I assumed that that was prompted by a change to Guava to remove its `checker-compat-qual` dependency, which presumably jimfs had been relying on. But:
- I think Guava had already removed that dep in an earlier release.
- jimfs does declare its own direct dep on `checker-compat-qual`.
- I had already had a CL out for the _real_ fix for the CI failure of https://github.com/google/jimfs/pull/240, which is related to modules. That said, I have no idea why the failure hasn't already been happening at head. Maybe it has something to do with a change in the version of Maven used by default in CI??? I'm not sure I believe that, but if that is the cause, then this is another reason to look at [Maven Wrapper](https://github.com/google/guava/pull/6326).

Anyway, this change _still_ migrates us, since `@NullableDecl` is discouraged nowadays and we can finally replace our usages with usages of type-use annotations.

This change is generated automatically with some tooling that we've already used elsewhere, including testing it on Guava, so I'd expect it to be accurate. For future reference, I'd note that migrations off `@NullableDecl` _could_ be tricky at scale because `@NullableDecl` means "This thing is null _unless it's an array (including varargs), in which case it's an array that can **contain** nulls_" and, to make things worse, the Checker Framework is (probably) the only tool that treats it properly.... Luckily, that doesn't appear to come up here.

RELNOTES=n/a
PiperOrigin-RevId: 536748545