Update Error Prone to 2.32.0 and only run it on JDK 17+.
Error Prone was actually not running on JDK 17+ at all due to profile activation only looking at the first of the JDK version ranges apparently. I don't know if that's a bug, but it seemed to be a valid version range specification according to https://maven.apache.org/enforcer/enforcer-rules/versionRanges.html.

The `-J` flags were also not actually being passed to the compiler, but that wasn't required on JDK 11, so nothing appeared to be broken. `fork` needs to be set to `true` to pass the `-J` flags. See
https://maven.apache.org/plugins/maven-compiler-plugin/compile-mojo.html#compilerArgs:
> Note that `-J` options are only passed through if `fork` is set to `true`.

RELNOTES=n/a
PiperOrigin-RevId: 678417880